feat: Complete Phase 4 - Consolidate processor components
