+++ src/fapilog/_internal/processors.py
@@ -2,397 +2,10 @@
 High-performance logging processors with enterprise-grade observability.
 """
 
-import re
-import time
-from collections import deque
-from typing import Any, Dict, List, Optional, Set, cast
-
-from .deduplication_processor import DeduplicationProcessor  # noqa: F401
-
 # Import other processors for re-export
-from .processor import Processor
-from .throttle_processor import ThrottleProcessor  # noqa: F401
-
-
-def _should_redact_at_level(event_level: str, redact_level: str) -> bool:
-    """Check if redaction should be applied based on log level."""
-    level_hierarchy = {"DEBUG": 0, "INFO": 1, "WARNING": 2, "ERROR": 3, "CRITICAL": 4}
-    event_priority = level_hierarchy.get(event_level.upper(), 1)
-    redact_priority = level_hierarchy.get(redact_level.upper(), 1)
-    return event_priority >= redact_priority
-
-
-class RedactionProcessor(Processor):
-    """
-    Enterprise-grade redaction processor with iterative algorithm for
-    optimal memory usage.
-
-    Optimized for real-world enterprise use cases:
-    - Memory-efficient O(n) iterative algorithm using explicit stack
-    - No stack overflow risk for deeply nested structures (2000+ levels)
-    - Built-in performance metrics and cache statistics
-    - Comprehensive observability for troubleshooting and monitoring
-    - Optional metrics disable for extreme performance scenarios (< 5% use cases)
-    - Circular reference detection and configurable depth limits
-
-    Performance: ~2-5ms for typical events (100-1000 fields) with full
-    observability.
-
-    The 15-20% overhead from metrics is acceptable for enterprise environments
-    where observability and troubleshooting capabilities are essential.
-    """
-
-    def __init__(
-        self,
-        patterns: Optional[List[str]] = None,
-        redact_level: str = "INFO",
-        enable_metrics: bool = True,
-        max_depth: int = 1000,
-        **config: Any,
-    ) -> None:
-        """Initialize enterprise-ready redaction processor with iterative
-        algorithm.
-
-        Args:
-            patterns: List of regex patterns to match for redaction
-            redact_level: Minimum log level for redaction to be applied
-            enable_metrics: Enable performance metrics and cache statistics
-                (recommended)
-            max_depth: Maximum nesting depth to prevent infinite loops
-                (default: 1000)
-            **config: Additional configuration parameters
-        """
-        self.patterns = patterns or []
-        self.redact_level = redact_level
-        self.enable_metrics = enable_metrics
-        self.max_depth = max_depth
-
-        # Pre-compiled patterns for performance
-        self.compiled_patterns: List[re.Pattern[str]] = []
-
-        # Enterprise observability features
-        if self.enable_metrics:
-            self.performance_stats = {
-                "total_operations": 0,
-                "total_time_ms": 0.0,
-                "avg_time_ms": 0.0,
-            }
-            self.pattern_cache: Dict[str, bool] = {}
-            self.cache_hits = 0
-            self.cache_misses = 0
-
-        super().__init__(
-            patterns=patterns, redact_level=redact_level, max_depth=max_depth, **config
-        )
-
-    async def _start_impl(self) -> None:
-        """Initialize processor with compiled patterns."""
-        self.compiled_patterns = [
-            re.compile(pattern, re.IGNORECASE) for pattern in self.patterns
-        ]
-
-    def validate_config(self) -> None:
-        """Validate configuration parameters."""
-        if not isinstance(self.patterns, list):
-            raise ValueError("patterns must be a list of strings")
-
-        for pattern in self.patterns:
-            if not isinstance(pattern, str):
-                raise ValueError("All patterns must be strings")
-            try:
-                re.compile(pattern)
-            except re.error as e:
-                raise ValueError(f"Invalid regex pattern '{pattern}': {e}") from e
-
-        if not isinstance(self.redact_level, str):
-            raise ValueError("redact_level must be a string")
-
-        if not isinstance(self.max_depth, int) or self.max_depth < 1:
-            raise ValueError("max_depth must be a positive integer")
-
-    def process(
-        self, logger: Any, method_name: str, event_dict: Dict[str, Any]
-    ) -> Dict[str, Any]:
-        """
-        Redact sensitive information with enterprise-grade performance and observability.
-        """
-        # Performance tracking for enterprise observability
-        start_time = time.perf_counter() if self.enable_metrics else None
-
-        try:
-            # Check if redaction should be applied based on log level
-            event_level = event_dict.get("level", "INFO")
-            if not _should_redact_at_level(event_level, self.redact_level):
-                return event_dict
-
-            if not self.patterns or not self.compiled_patterns:
-                return event_dict
-
-            # Use memory-efficient iterative redaction with safety features
-            return self._redact_with_observability(event_dict)
-
-        finally:
-            # Track performance metrics for enterprise monitoring
-            if self.enable_metrics and start_time is not None:
-                elapsed = (time.perf_counter() - start_time) * 1000
-                self.performance_stats["total_operations"] += 1
-                self.performance_stats["total_time_ms"] += elapsed
-                ops = self.performance_stats["total_operations"]
-                self.performance_stats["avg_time_ms"] = (
-                    self.performance_stats["total_time_ms"] / ops
-                )
-
-    def _redact_with_observability(self, event_dict: Dict[str, Any]) -> Dict[str, Any]:
-        """Memory-efficient iterative redaction with safety features and in-place modification."""
-
-        # Pre-check depth to avoid processing overly complex structures
-        if self._estimate_depth(event_dict) > self.max_depth:
-            raise ValueError(f"Object nesting exceeds maximum depth: {self.max_depth}")
-
-        # Iterative redaction using explicit stack for O(n) memory complexity
-        return cast(Dict[str, Any], self._redact_iterative(event_dict))
-
-    def _estimate_depth(self, obj: Any, max_check_depth: int = None) -> int:
-        """Estimate maximum nesting depth of an object efficiently.
-
-        Args:
-            obj: Object to analyze
-            max_check_depth: Maximum depth to check before giving up (defaults to max_depth + 50)
-
-        Returns:
-            Estimated depth (capped at max_check_depth)
-        """
-        if not isinstance(obj, (dict, list)):
-            return 0
-
-        # Use a reasonable limit for depth checking
-        if max_check_depth is None:
-            max_check_depth = self.max_depth + 50
-
-        # Use iterative approach for depth estimation too
-        stack = deque([(obj, 0)])  # (object, current_depth)
-        max_depth = 0
-
-        while stack and max_depth < max_check_depth:
-            current, depth = stack.popleft()
-            max_depth = max(max_depth, depth)
-
-            if isinstance(current, dict):
-                for value in current.values():
-                    if isinstance(value, (dict, list)) and depth < max_check_depth:
-                        stack.append((value, depth + 1))
-            elif isinstance(current, list):
-                for item in current:
-                    if isinstance(item, (dict, list)) and depth < max_check_depth:
-                        stack.append((item, depth + 1))
-
-        return max_depth
-
-    def _redact_iterative(self, obj: Any) -> Any:
-        """Iterative redaction using explicit stack instead of recursion.
-
-        This eliminates O(n²) memory complexity and stack overflow issues.
-        Uses depth-first traversal for consistent ordering.
-
-        Args:
-            obj: Root object to redact (modified in-place)
-
-        Returns:
-            The redacted object (same as input, modified in-place)
-        """
-        # Handle None and non-container types early
-        if obj is None or not isinstance(obj, (dict, list)):
-            return obj
-
-        # Stack contains: (current_obj, parent_obj, key_or_index, path, depth)
-        stack = deque([(obj, None, None, [], 0)])
-        visited: Set[int] = set()  # Circular reference detection
-
-        while stack:
-            current, parent, key, path, depth = stack.pop()  # DFS: use pop() for LIFO
-
-            # Safety check: prevent infinite loops with depth limit
-            if depth > self.max_depth:
-                continue
-
-            # Circular reference detection
-            obj_id = id(current)
-            if obj_id in visited:
-                continue
-
-            # Only track objects that can contain references (dict/list)
-            if isinstance(current, (dict, list)):
-                visited.add(obj_id)
-
-            if isinstance(current, dict):
-                # Process dictionary items
-                # We need to collect items first to avoid modifying dict during iteration
-                items = list(current.items())
-
-                for k, v in items:
-                    current_path = path + [k]
-
-                    # Check if key matches any pattern
-                    key_matches = self._pattern_matches(str(k))
-
-                    if isinstance(v, str):
-                        # Handle string values
-                        value_matches = self._pattern_matches(v)
-                        if key_matches or value_matches:
-                            current[k] = "[REDACTED]"
-                    elif isinstance(v, (dict, list)):
-                        # Schedule nested structures for processing
-                        if depth < self.max_depth:
-                            stack.append((v, current, k, current_path, depth + 1))
-                        # If key matches, mark for redaction after processing children
-                        if key_matches:
-                            # We'll redact this after processing children
-                            # Add a marker to redact this key later
-                            stack.append(
-                                ("__REDACT_KEY__", current, k, current_path, depth)
-                            )
-                    else:
-                        # For other types (int, float, bool, None), redact if key matches
-                        if key_matches:
-                            current[k] = "[REDACTED]"
-
-            elif isinstance(current, list):
-                # Process list items
-                for i, item in enumerate(current):
-                    current_path = path + [i]
-
-                    if isinstance(item, (dict, list)):
-                        # Schedule nested structures for processing
-                        if depth < self.max_depth:
-                            stack.append((item, current, i, current_path, depth + 1))
-                    elif isinstance(item, str):
-                        # Check if string value matches patterns
-                        if self._pattern_matches(item):
-                            current[i] = "[REDACTED]"
-
-            elif current == "__REDACT_KEY__":
-                # Special marker to redact a key after its children have been processed
-                if parent is not None and key is not None:
-                    parent[key] = "[REDACTED]"
-
-        return obj
-
-    def _pattern_matches(self, text: str) -> bool:
-        """Check if text matches any pattern with optional caching for enterprise observability."""
-        if not self.enable_metrics:
-            # Fast path without caching for extreme performance scenarios
-            return any(pattern.search(text) for pattern in self.compiled_patterns)
-
-        # Enterprise path with caching and metrics
-        if text in self.pattern_cache:
-            self.cache_hits += 1
-            return self.pattern_cache[text]
-
-        result = any(pattern.search(text) for pattern in self.compiled_patterns)
-
-        # Simple cache management - clear when too large
-        if len(self.pattern_cache) > 1000:
-            self.pattern_cache.clear()
-
-        self.pattern_cache[text] = result
-        self.cache_misses += 1
-        return result
-
-    def get_performance_metrics(self) -> Dict[str, Any]:  # noqa  # vulture: ignore
-        """Get comprehensive performance metrics for enterprise monitoring."""
-        if not self.enable_metrics:
-            return {
-                "metrics_enabled": False,
-                "note": "Enable metrics for observability",
-            }
-
-        metrics = self.performance_stats.copy()
-        metrics["metrics_enabled"] = True
-
-        # Cache statistics
-        total_cache_ops = self.cache_hits + self.cache_misses
-        hit_rate = self.cache_hits / max(1, total_cache_ops)
-
-        metrics.update(
-            {
-                "cache_hits": self.cache_hits,
-                "cache_misses": self.cache_misses,
-                "cache_hit_rate": hit_rate,
-                "cache_size": len(self.pattern_cache)
-                if hasattr(self, "pattern_cache")
-                else 0,
-            }
-        )
-
-        return metrics
-
-    def clear_cache(self) -> None:  # vulture: ignore
-        """Clear pattern matching cache for enterprise cache management."""
-        if self.enable_metrics and hasattr(self, "pattern_cache"):
-            self.pattern_cache.clear()
-
-    def reset_metrics(self) -> None:  # vulture: ignore
-        """Reset all performance metrics for enterprise monitoring."""
-        if self.enable_metrics:
-            self.performance_stats = {
-                "total_operations": 0,
-                "total_time_ms": 0.0,
-                "avg_time_ms": 0.0,
-            }
-            self.cache_hits = 0
-            self.cache_misses = 0
-
-
-class FilterNoneProcessor(Processor):
-    """Processor that filters out None events."""
-
-    def __init__(self, **config: Any) -> None:
-        """Initialize filter processor.
-
-        Args:
-            **config: Configuration parameters (not used by this processor)
-        """
-        super().__init__(**config)
-
-    def process(
-        self, logger: Any, method_name: str, event_dict: Optional[Dict[str, Any]]
-    ) -> Optional[Dict[str, Any]]:
-        """Filter out None events."""
-        if event_dict is None:
-            return None
-        return event_dict
-
-
-class SamplingProcessor(Processor):
-    """Processor that drops events probabilistically for sampling."""
-
-    def __init__(self, rate: float = 1.0, **config: Any) -> None:
-        """Initialize sampling processor.
-
-        Args:
-            rate: Sampling rate between 0.0 and 1.0
-            **config: Additional configuration parameters
-        """
-        self.rate = rate
-        super().__init__(rate=rate, **config)
-
-    def validate_config(self) -> None:
-        """Validate configuration parameters."""
-        if not isinstance(self.rate, (int, float)):
-            raise ValueError("rate must be a number")
-
-        if not 0.0 <= self.rate <= 1.0:
-            raise ValueError("rate must be between 0.0 and 1.0")
-
-    def process(
-        self, logger: Any, method_name: str, event_dict: Dict[str, Any]
-    ) -> Optional[Dict[str, Any]]:
-        """Sample log events based on the configured rate."""
-        import random
-
-        if self.rate >= 1.0:
-            return event_dict
-
-        if random.random() >= self.rate:
-            return None  # Drop this event
-        return event_dict
+from ..processors.base import Processor  # noqa: F401
+from ..processors.deduplication import DeduplicationProcessor  # noqa: F401
+from ..processors.filtering import FilterNoneProcessor  # noqa: F401
+from ..processors.redaction import RedactionProcessor  # noqa: F401
+from ..processors.sampling import SamplingProcessor  # noqa: F401
+from ..processors.throttling import ThrottleProcessor  # noqa: F401
