+++ src/fapilog/_internal/templates/processor_template.py
@@ -18,7 +18,7 @@ import time
 from typing import Any, Callable, Dict, Optional, cast
 
 from ...exceptions import ProcessorConfigurationError
-from ..async_processor_base import AsyncProcessorBase
+from ...processors.async_base import AsyncProcessorBase
 
 logger = logging.getLogger(__name__)
