+++ src/fapilog/_internal/testing/processor_testing.py
@@ -11,7 +11,7 @@ from typing import Any, Dict, cast
 
 import pytest
 
-from ..async_processor_base import AsyncProcessorBase
+from ...processors.async_base import AsyncProcessorBase
 from ..processor import Processor
