+++ src/fapilog/core/registries/processor_registry.py
@@ -6,7 +6,7 @@ register custom processor implementations for use in the logging pipeline.
 
 from typing import Callable, Dict, Optional, Type
 
-from ..._internal.processor import Processor
+from ...processors.base import Processor
 
 
 class ProcessorRegistry:
