+++ src/fapilog/processors/async_base.py
@@ -12,7 +12,7 @@ from typing import Any, Awaitable, Callable, Dict, Optional
 
 from ..async_components.cache.safe_async_cache import SafeAsyncCache
 from ..async_components.concurrency.lock_manager import ProcessorLockManager
-from .processor import Processor
+from ..processors.base import Processor
 
 logger = logging.getLogger(__name__)
