+++ src/fapilog/processors/base.py
