+++ src/fapilog/testing/debug.py
@@ -5,8 +5,8 @@ import inspect
 import traceback
 from typing import Any, Dict, List, Type
 
-from .._internal.sink_factory import create_custom_sink_from_uri
-from .._internal.sink_registry import SinkRegistry
+from ..core.factories.sink_factory import create_custom_sink_from_uri
+from ..core.registries.sink_registry import SinkRegistry
 from ..sinks import Sink
