+++ tests/test_processor_registry.py
@@ -5,8 +5,8 @@ from typing import Any, Dict, Optional, Type
 
 import pytest
 
-from fapilog._internal.processor import Processor
 from fapilog._internal.processor_registry import ProcessorRegistry, register_processor
+from fapilog.processors.base import Processor
 
 
 # Test processor implementations
